refactor: appbar 컴포넌트 분리
